Uncomment methods for returning instance of the repo and data source
- The purpose of InjectorUtils is to provide static methods for
dependency injection.

- Dependency injection is the idea that you should make required
components available for a class, instead of creating them within the
class itself. An example of how the Sunshine code does this is that
instead of constructing the WeatherNetworkDatasource within the
SunshineRepository, the WeatherNetworkDatasource is created via
InjectorUtilis and passed into the SunshineRepository constructor. One
of the benefits of this is that components are easier to replace when
you're testing
Signed-off-by: Alvin Dizon <alvin.dizon@ghl.com>